Move GenerateCommandSuccessOutput
